fix: return actual PV size on allocation (#116)
* fix: return actual PV size on allocation

LVM always aligns the size of the LV to default 4MiB size boundary.
If the requested size did not match to 4Mi, its extended to match it.
But the PV size returned still showed the user requested size. This PR
gets the actual size of the LV post its allocation and this is what is
now returned.

* fix: incorporate review comments

* fix: return volume size when EnsureVolume finds a volume